move babel file filtering into core
This adds a new little bit of API that stage3 packagers need to know about, but I think it ultimately simplifies them by doing more of the work in core. And it allows us to have consistent rules for which files go through babel.
