+++ packages/core/src/index.ts
@@ -16,3 +16,4 @@ export { compile as jsHandlebarsCompile } from './js-handlebars';
 export { AppAdapter, AppBuilder, EmberENV } from './app';
 export { todo, unsupported, warn, debug, expectWarning, throwOnWarnings } from './messages';
 export { explicitRelative } from './paths';
+export { default as babelFilter } from './babel-filter';
