+++ packages/core/src/to-broccoli-plugin.ts
@@ -1,7 +1,6 @@
 import Plugin from 'broccoli-plugin';
 import { Packager, PackagerInstance } from './packager';
 import Stage from './stage';
-import PackageCache from './package-cache';
 
 interface BroccoliPackager<Options> {
   new (stage: Stage, options?: Options): Plugin;
@@ -21,24 +20,12 @@ export default function toBroccoliPlugin<Options>(packagerClass: Packager<Option
     async build() {
       if (!this.packager) {
         let { outputPath, packageCache } = await this.stage.ready();
-        // stages are allowed to share a package cache as an optimization, but
-        // they aren't required to. Whereas Packagers are allowed to assume they
-        // will receive a packageCache instance.
-        //
-        // We also always register a shared stage3 packageCache so it can be
-        // used by things like babel plugins and template compilers.
+        // We always register a shared stage3 packageCache so it can be used by
+        // things like babel plugins and template compilers.
         if (packageCache) {
           packageCache.shareAs('embroider-stage3');
-        } else {
-          packageCache = PackageCache.shared('embroider-stage3');
         }
-        this.packager = new packagerClass(
-          outputPath,
-          this.outputPath,
-          msg => console.log(msg),
-          packageCache,
-          this.options
-        );
+        this.packager = new packagerClass(outputPath, this.outputPath, msg => console.log(msg), this.options);
       }
       return this.packager.build();
     }
